+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## [1.19.4-1.0.0.beta.6] - Unreleased
 
 ### Added missing recipe for tool crafting table
+### Added all tool set for ruby
+### Added aquarium trident.
+### Added shields for all custom ores.
 
 ## [1.19.4-1.0.0.beta.5] - 2023-5-3